extract search as separate component
